Move Idenity and ExitIdentity abi encoding logic into althea_types crate
This patch moves the ethereum abi encoding logic for Identity and
ExitIdentity into the althea_types crate. This is a more logical
location for the code and in the process of moving it the tests have
also been improved in scope and quality, resolving some logic issues
that existed in the previous version.
