+++ althea_types/Cargo.toml
@@ -21,7 +21,7 @@ lettre = { version = "0.10", features = ["serde"] }
 ipnetwork = "0.20"
 bincode = "1.3"
 deep_space = { workspace = true }
+rand = "0.8"
 
 [dev-dependencies]
-rand = "0.8"
 sodiumoxide = "0.2"
