+++ althea_types/src/error.rs
@@ -6,12 +6,16 @@ use std::fmt::Result as FormatResult;
 #[derive(Clone, Debug)]
 pub enum AltheaTypesError {
     WgParseError(DecodeError),
+    BadEthAbiInput(String),
 }
 
 impl fmt::Display for AltheaTypesError {
     fn fmt(&self, f: &mut fmt::Formatter) -> FormatResult {
         match self {
             AltheaTypesError::WgParseError(val) => write!(f, "Failed to parse WgKey with {val}"),
+            AltheaTypesError::BadEthAbiInput(e) => {
+                write!(f, "Failed to parse Eth ABI input with {e}")
+            }
         }
     }
 }
